Extract tournaments handlers to own folder and module
