+++ src/tournaments/index.js
@@ -1,0 +1,4 @@
+
+module.exports = {
+    handlers: require('./handlers'),
+};
